Remove lock from AggregateDependenciesSnapshotProvider
The collection protected by this lock changes infrequently, yet it is
read from very regularly during dependency updates. It's a good
fit for opportunistic interlocked operations on an immutable dictionary.

The lock also introduced the potential for contention across projects,
as this AggregateDependenciesSnapshotProvider brokers data from
all projects in the solution.